feat(palantir): add full logic in scrubber (without connecting with storage)

+++ marline_palantir/src/types/super_feature.rs
@@ -1,18 +1,18 @@
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct SuperFeature {
     tier_id: u8,
-    value: u32,
+    value: u64,
     version_id: u32,
 }
 
 impl SuperFeature {
-    pub fn new(tier_id: u8, value: u32, version_id: u32) -> Self {
+    pub fn new(tier_id: u8, value: u64, version_id: u32) -> Self {
         Self { tier_id, value, version_id }
     }
     pub fn tier_id(&self) -> u8 {
         self.tier_id
     }
-    pub fn value(&self) -> u32 {
+    pub fn value(&self) -> u64 {
         self.value
     }
